requirements.txt
streamlit==1.18.1
python-dotenv==1.0.0
requests==2.28.2
langchain_openai==0.0.1
openai==0.27.0
google-generativeai



